requirements.txt
# AI/ML Module Dependencies for Public Speaking Coach Web App

# Speech-to-Text Transcription
faster-whisper>=0.10.0

# Google ADK for NLP Analysis
google-generativeai>=0.3.0

# FastAPI and async support
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Firebase Integration
firebase-admin>=7.1.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# General utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Optional: For enhanced audio support
pydub>=0.25.0
ffmpeg-python>=0.2.0
